Реализует методы сервиса оборота товаров

+++ hw19-final/back/src/main/java/ru/otus/hw/models/dto/OrderDto.java
@@ -3,6 +3,7 @@ package ru.otus.hw.models.dto;
 import lombok.Data;
 import ru.otus.hw.models.enums.OrderStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Set;
 import java.util.UUID;
@@ -18,7 +19,7 @@ public class OrderDto {
 
     private OrderStatus status;
 
-    private double totalPrice;
+    private BigDecimal totalPrice;
 
     private UUID userId;
